service and controller

+++ src/main/java/com/example/crud/entity/studententity.kt
@@ -3,7 +3,6 @@ package com.example.crud.entity
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import jakarta.persistence.Table
-import org.springframework.context.annotation.Primary
 
 
 @Entity//managed by spring JPA not by sping IOC container
@@ -14,14 +13,14 @@ data class Student
 
     (
     @Id
-    var id: Long = 0,//to tell primary key
+    var id: Int = 0,//to tell primary key
 
 
     var name: String ="name1",
     var age: Int = 22,
-    var mail :  String = "name@mail.com",
+    var mail: String = "name@mail.com",
     var subject: String = "none",
-    var deleted: Boolean  = false,
+    var deleted: Boolean = false,
 
 
     ) {
